requirements.txt
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1
groq==0.25.0
requests==2.32.3
python-dotenv==1.0.1